Changed queries structure
Changed queries structure
